delete/change type cast (int => uint64)

+++ handlers/reset_password.go
@@ -104,7 +104,7 @@ func ChangePassword(c echo.Context) error {
 	}
 	//check Borrower ID
 	borrowerModel := models.Borrower{}
-	err = borrowerModel.FindbyID(int(uuid_reset_password.Borrower.Int64))
+	err = borrowerModel.FindbyID(uint64(uuid_reset_password.Borrower.Int64))
 	if err != nil {
 		return returnInvalidResponse(http.StatusForbidden, err, "Akun Tidak ditemukan")
 	}
